Create README.md
Background:
Our client, who had previously worked with us for his real estate business, approached us with a new project. He wanted us to scrape data from the popular property listing website, Zillow. He needed a solution that would allow him to enter specific filters, such as zip code, price range, and property type, to obtain a list of properties with relevant information, including address, actual price, zestimate, and other crucial attributes for prospective buyers. He was particularly interested in identifying properties being sold in distress, where the zestimate was higher than the actual price.

Problem:
We expressed our concern to the client about using web scraping to acquire data, as it could be considered unethical to obtain someone else's data for free, and a slight change in site layout could make the code obsolete. Moreover, most modern sites are designed to prevent bots from scraping data.

Solution:
Despite our concerns, the client was ready to take the risk and explore the feasibility of the project. As expected, our initial attempt to scrape the data was blocked by Zillow. After extensive research, we discovered a third-party API that could extract the required data. This was a breakthrough, as even when the website layout changed, the API could be updated to maintain code functionality.

We shared our findings with the client and presented the pricing details for purchasing the API. After confirming with the client, we selected 30 columns from the available 256 attributes for each property and loaded the data onto a Google Sheet, allowing easy access from any device. We used Google Colab as the IDE since we knew it was the client's preferred tool and provided training on how to modify property filters. The client was pleased with the outcome.